Device Support: Add canRestore() to device factory.
Also get rid of supportsDeviceType(), which is no longer needed now that
most services live in IDevice itself.
The new approach is more flexible as well, allowing factories to refuse
loading a device even if they know its type (e.g. because it is no
longer relevant).
We also rename loadDevice() to restore(), because that is more
exact and is also the name that is being used elsewhere in QtCreator.

Change-Id: I51f57c118a048c1a781a55a94966d554d7aed7a2
Reviewed-by: Tobias Hunger <tobias.hunger@nokia.com>