Qt4Nodes: Fix wrong "Parsing in progress" warning for some cases
Removing a subproject and changing the project type of a parent project
didn't set always the valid parse state nor emitted the parsing finished
signal. Also ensure that removing and then readding a node does update
the runconfiguration for that node.

Note: The error message is now: "The .pro file %1 could not be parsed",
which is better, but can be improved, improving that is QTCREATORBUG-7377

Task-number: QTCREATORBUG-7394

Change-Id: Ie0d9a430ee902bd194179b0512ac8295bc519f2a
Reviewed-by: Daniel Teske <daniel.teske@nokia.com>