Fix broken handling of unsupported mime types.
Task-number: QTCREATORBUG-2701
Task-number: QTCREATORBUG-4059

Enable plugins to support pdf, etc. files out of the box in
Qt Creator. If we still fail we try to open the link using
the desktop service, as last resort to open a page showing
some error message and hints about the error. Adjust the old
page not found message to match the new design.

Change-Id: I7fef79d27e8bc279a6169b6f3da4787ffa0d023d
Reviewed-by: Niels Weber <niels.2.weber@nokia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@nokia.com>
Reviewed-by: Karsten Heimrich <karsten.heimrich@nokia.com>