C++: Change back highlighting of types
The type highlighting change part of the recent patch
4a2a17af8aa157cadb4ec230058d9ae34e11fd66 didn't seem to
please much from a visual point of view. It's a better
idea to keep the type highlighting as it was for now
and in the future try again the approach with an explicit
option and perhaps a more restrictive context.

The other patch is not reverted entirely because it does
fix a couple of bugs.

Change-Id: I806afa3d8c1c4b241080b8704255d737f61ee12c
Reviewed-by: Roberto Raggi <roberto.raggi@nokia.com>