Making the QmlApplicationViewer Qt5 compatible
This patch makes a "QmlApplicationViewer" based application from
the Qt Quick app wizard buildable and runnable with Qt5. Note,
that this is still QtQuick 1.1. We are still missing a QtQuick 2.0
based pendant of QmlApplicationViewer.

In addition (separate patch), we may want to tweak the wizard
flags in a way that the wizard allows to create a "Qt Quick 1"
application for a Qt5 target. (Or perhaps we explicitly do not
want that, because we want the developer to use QtQuick 2 with
Qt5?)

Change-Id: Ie3fc8a77a1ff4fba3244f2047d2685c570014356
Reviewed-by: Eike Ziller <eike.ziller@nokia.com>
Reviewed-by: Alessandro Portale <alessandro.portale@nokia.com>