Provide build system independent deployment data.
Attach an object containing deployment information
to the target. The respective build system plugins are
responsible for setting up this information. The
Qt4ProjectManager already does that; others should
follow (which is expected to result in gradual
refinements of the design).
The immediate motivation here is to free the
RemoteLinux plugin from the requirement that
qmake must be used for building the projects,
which makes it much less generic than it could
and should be.

Change-Id: I9943787f4e352d014e721082016542b10c8cce90
Reviewed-by: Daniel Teske <daniel.teske@nokia.com>