C++: Fix regression when binding qualified names
This was introduced in function extraction refactoring intended
to fix an issue with invalid class names: a66e344b424643fce71c127fdfb0a21103ca0171
The patch fixes only the regression itself - the previous fix
is correct.

The report below consists two parts. The other one is not a
regression but nevertheless is fixed by the previous patch.

Task-number: QTCREATORBUG-7730
Change-Id: I6f65584902619b542c9ce56cd0f37218a3d50104
Reviewed-by: Roberto Raggi <roberto.raggi@nokia.com>