Introduce static localProcesses() method to DeviceProcessList.
Add synchronous static getLocalProcesses() (platformspecific)
method to LocalProcessList, use that for the model, simplifying
the code of doUpdate(). Use that for
DeviceProcessList::localProcesses().

Polish the proclist-method for UNIX a bit such that it tries
to read out the 'cmdline' first to obtain the complete
command line and full path to the executable.

Similarly, set the absolute image path on Windows.

Implement doKill() on Windows.

Change-Id: I3a5b990eaa9f6beb4c589406bdf4c7a234e40734
Reviewed-by: hjk <qthjk@ovi.com>