Fix running applications in non-xterm terminals
Setting Tools/Options/Environment/General/Terminal to one of the
following works now as expected.

  1. xterm -e
  2. aterm -e
  3. rxvt -e
  4. urxvt -e
  5. xfce4-terminal -x
  6. konsole --nofork -e
  7. gnome-terminal -x

Without the '--nofork' option for konsole, from creator's pov, the
terminal emulator will finish immediately and therefore creator closes
the local socket. As a result of this, the user is presented with sth.
like this (from the stub):

    Cannot connect creator comm socket /tmp/qt_temp.A13523/stub-socket:
    No such file or directory.

Task-Number: QTCREATORBUG-1633
Change-Id: I732e20d34aebdc1591355e0c4e78c907f439924c
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>