Debugger: Improve engine detection.
Remove the RunConfiguration parameter from guessCppEngineType()
and use the DebuggerStartParameter::toolChainAbi only.

Use fillParameter() to set the debugger command and toolchain
in order to prevent mismatches.

Add logic trying to figure out the profile to fillParameters()
in case the profile passed in is 0 for the command line cases.

Use CDB matcher for post mortem and Remote CDB, fix it to
prefer 64bit CDB on 64bit OS.

Change-Id: Icedc3883fe15660303060498ab609957e6d01cd1
Reviewed-by: hjk <qthjk@ovi.com>