Introduce $$QTC_PREFIX instead of abusing $(INSTALL_ROOT).
Reason: $INSTALL_ROOT is only meant to be used by packagers
to temporarily put the contents into a different location,
which is needed for fakeroot packaging.

QTC_PREFIX is not a qmake variable, and defaults to
/usr/local. On Windows the default prefix is "QtCreator",
since "make install" is expected to be used in
a packaging context only where either INSTALL_ROOT
should be used or QTC_PREFIX should be set to the
absolute destination path where e.g. an installer
generator will pick it up.

Change-Id: Ifa4950340e58e34726c53f5417adcc7b50828ce1
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>