Exclude the qml designer from the qbs build.
The qbs build of this plugin is constantly being broken by its
maintainers. Due to its convoluted structure, it's hard for others to
keep the qbs files up to date, so we give up on that for now.
We can re-enable it later if and when there is more consideration
for qbs.
(This patch also contains one last build fix.)

Change-Id: I37f8bea3ee65cb62857e7d6649eab4e6d129eb5b
Reviewed-by: Marco Bubke <marco.bubke@digia.com>