Import to all possible kits
List any detected build in all kits that match the characteristics
of the project build settings found when importing a build for
unconfigured projects.

So far we just listed it with the first kit we found which is
surprising sometimes (e.g. when having a embedded linux and a desktop
kit using the same Qt version).

Change-Id: I9d1de26eca3dcabd424064a66019f5f6be876249
Reviewed-by: Daniel Teske <daniel.teske@digia.com>