CrashHandler: Add "Attach and Debug" button.
This will launch a new instance of Qt Creator attaching to the crashed
instance.

The 'Restart' functionality is now represented as a check box since it
would be confusing to have one button that restarts the app and quits
the crash handler and another one that starts a debugger but keeps the
crash handler open (which is necessary, otherwise the crashed app will
quit).

Change-Id: Id88f418ff73ab7bc72b05753ce2b61bbef8f30cf
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>