Help: Prevent moc from running on helpviewer_qwv.cpp.
If Webkit is disabled, moc will run on the file anyway, because there is
a ".moc" included and moc does not know about #ifdefs. This results in
an ugly warning ("no output generated") being emitted in the no-Webkit
case.
Solve the problem by moving the declaration of the QObject subclass into
the header file.

Change-Id: I2abaffd85a342fa2f3dbaa7bcf3ef63a2972c73e
Reviewed-by: Karsten Heimrich <karsten.heimrich@digia.com>
Reviewed-by: hjk <qthjk@ovi.com>