Device support: Ensure UI changes are always propagated to the device.
The original problem was that we cannot assume
QLineEdit::editingFinished() is emitted when clicking a button while a
QLineEdit has the focus (QTCREATORBUG-1675). The original solution to
that introduced two more bugs, namely that changes are not visible in
other tabs until the dialog has been closed and reopened and that
pressing "Apply" also applies changes that are done later and which the
user intends to revert by by pressing "Cancel" (QTCREATORBUG-7288).
This patch intends to fix all of these issues.

Task-number: QTCREATORBUG-1675
Task-number: QTCREATORBUG-7288
Change-Id: I569a89f64843e08ee389e3eba6bdcb473ba22393
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>