C++ Macro Usages: Fix search result encoding and highlighting.
If the result lines contain non-Latin1 characters (e.g. in comments),
the output is garbled. Using QByteArrays with toUtf8() and fromUtf8()
doesn't seem to be an alternative, as in this case the macro.offset() is
no longer valid.

So it seems to be the best to use QString for the result lines.

Task-number: QTCREATORBUG-7122
Change-Id: I57128c9c9f3eb182f079e305e97e9c5ac0a1bc61
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>