Mac: Don't link to 10.7+ api even when compiling there.
The 'fullscreen' feature uses constants that are only available
on 10.7 or later, but when compiling on 10.6 we just defined the
constants ourselves.
With this patch we define the constants ourselves even on 10.7 and
later, to avoid linking to 10.7-only symbols when compiling on 10.7
or later, making it possible to run a Qt Creator compiled against
a 10.7 SDK also on 10.6.

Change-Id: I65236a85bdda3473fab0e424270b96880943d685
Reviewed-by: Erik Verbruggen <erik.verbruggen@digia.com>
Reviewed-by: Tim Jenssen <tim.jenssen@digia.com>