Qt: Allow for custom environments to run "qmake -query" in
Some SDKs (Blackberry) use one qmake only and have environment
variables embedded in their mkspecs that are then used to switch
between different instances of that SDK. Rather strange approach,
but that is what they do.

So this patch provides a member method to BaseQtVersion that can
be used to override the environment seen by "qmake -query". The
default implementation is to use the system environment.

This will not work very well while trying to initially detect which
kind of Qt version to construct for a qmake. This is not a big issue
though as they can either create the Qt version in non-standard ways
(Blackberry is going for a SDK setup page). If the system environment
is set up correctly the normal add-a-qmake approach will also work.

Change-Id: I80b1c259b127de816b5d02da414a9f4b11d7f658
Reviewed-by: Mehdi Fekari <mfekari@rim.com>
Reviewed-by: Daniel Teske <daniel.teske@digia.com>