Inspector: Disable object selection from Editor
Clicking around QML code in the editor would update
the Inspector. This is now disabled.

The problem with the behaviour is that the object tree
would be expanded and that may be unwelcome for some users.
Also some code snippets may be shared by multiple
objects, for eg. a custom QML component. In this case, the
behaviour was undefined as to which object in the tree would
be selected. This case of 'one-to-many' was not handled. Finally,
the inspector queries for object ids lazily. Hence, sections of
code may not be associated with an id resulting in unknown
behaviour.

Task-number: QTCREATORBUG-8246

Change-Id: I6266fd0767d0619af70387e2e867a756548120ca
Reviewed-by: Kai Koehne <kai.koehne@digia.com>