Fix building against Qt5 on MacOS.
Qt5 will set the mmacosx-version-min to either 10.6 or 10.7, depending
on the mkspec used. These lines resulted in two mmacosx-version-min
flags, when libc++ is used (macx-clang-libc++ mkspec), this leads to
compiler errors. As Qt/qmake already adds the flag, there should not be
a need to set it in the .pri/.pro files.

Change-Id: Ib8cc59d2b064737aa0b5a889b2b3183eceaea1f3
Reviewed-by: Eike Ziller <eike.ziller@digia.com>