Fix Qt Creator manual style when generating with Qt 5.
The HTML.templatedir is no longer used in Qt 5. Instead, the files
are resolved relative to the referencing qdocconf.
So we set the templatedir to the path of the qdocconf (for Qt 4)
and change the references to be relative to that path.
Also, it's HTML.extraimages in Qt 5, instead of extraimages.HTML

Task-number: QTBUG-28304
Change-Id: I5b253ea9fac5825c49f74460df188a7fca5c89cd
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@digia.com>