FakeVim: Focus window in given direction
Command <c-w><right> focuses the window to the right from the right edge
of the current window (not on the right side of the text cursor).
Similar behavior applies on the other directions.

For example, with layout (* is current window/split):
|   1*  |
| 2 | 3 |
on <c-w><right> the window number 3 should never be current.

Change-Id: I658011ef597bcadf4a77065ed0b31133ab72aa06
Reviewed-by: hjk <hjk121@nokiamail.com>