QmlDesigner: crash fix in document handling
If in case of an error we should not detach the rewriter
and we always should attach the other views for simplicity.

The model is a valid ( Item {} ) and all the views are in a proper state.

This makes recovering and changing the mode back simpler, since the error
case is not a special one.

We do disable the widgets and show the error message of course.

Change-Id: I433a31a3c9ae36909089dff4d1eb1716141e809b
Reviewed-by: Marco Bubke <marco.bubke@digia.com>