Do not clean whitespaces/ensure final line in autosave
Calling beginEditBlock() triggers a textChanged() signal and
did "confuse" Qt Quick Designer.

We could just suppress the signal to work around the issue, but I think
it is cleaner to not alter the text document in autosave at all.

Changing the current text document in autosave might have other
unforeseeable implications.

What happened in the Qt Quick Designer was that textChanged() triggered
a reload of the document.
While the reload of the document usually just triggered a
short/unnoticed dealay, it can also trigger a message box with warnings.

Since we (correctly) assume the .QML file was altered outside
of Qt Quick Designer we do sanity checks and warn about code that
might not be supported.

Having those warnings whenever a file is autosaved is annoying.

Change-Id: I62f5486878956cbb776998e4aba468a91cb9ec8e
Reviewed-by: hjk <hjk121@nokiamail.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>