DeviceManager: Read devices from global settings path
The global settings path is usually used by the SDK. The devices
listed in the global settings path are read first. A version field
has been introduced to ensure that the device with higher version is
stored in the model in case there is a clash between devices listed
in global settings and user settings.

Change-Id: I6472a3ee1f249bba1ac1a3ffe972a140c2a40ecd
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>