Fix Android plugin to support Qt 5 style rules-files
In Qt 5 we have modularized the rules files so that each module
specifies its own rules. This patch updates the plugin to support
reading from these in addition to the rules.xml of Qt 4 so that
it should work well for both.

The "replaces" attribute of the lib element is also added in Qt 5
because the files are generated and thus the approach of rules.xml
where replaces is a separate element and where the order has
semantics was not convenient. We of course support the
Necessitas approach as well.

Change-Id: Ife21903c9faf61e1791cf8c0ea78bb61c74dcb34
Reviewed-by: Daniel Teske <daniel.teske@digia.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>