QmlDesigner.NodeInstanceView: introducing delay in setPathToQt()
We have to delayed the restart, because when creating a
new project we switch to that project before we open
the new file. This means the user can get an error about
a missing puppet, because the new project is based on another Qt version.

This is a hotfix for Qt Creator 2.7.

Task-number: QTCREATORBUG-8756
Change-Id: I8c74c6334dca35ad29771339855e777d307b9e43
Reviewed-by: Marco Bubke <marco.bubke@digia.com>