DeviceManager: Delay loading of devices until extensions are initialized
The DeviceManager should be available after the projectexplorer plugin
was initialized so that other objects can connect to it, but it should only
load devices once all factories are initialized (in extensionsInitialized).

So split up construction of the object and loading.

Also remove list of inactive autodetected devices. It is unused and makes
it impossible to use addDevice(...) for loading.

Change-Id: I38e4ea9da9277d8d75476035c4763dc60b1472c8
Reviewed-by: Christian Kandeler <christian.kandeler@digia.com>