Device support: Allow interactive removal of auto-detected devices.
Now that there is no such thing as "inactive" devices anymore, users
need a way to get rid of devices that were at some point auto-detected
but are no longer present and won't be needed anymore. This is achieved
by making auto-detected devices removable if (and only if) they are
currently in the disconnected state.

Change-Id: Ibcdbf93170507550b6483aa33f428d2504baceb7
Reviewed-by: Tobias Hunger <tobias.hunger@digia.com>